Commit Summary. -John Rendell
-Add a parameter on route modify profile for fetching the image base64.
-Add a limit on parser at backend.
-Disable current modal and modal open boolean on retry connection button when socket got disconnected. So when player reconnect, they can move again, this is when the connection cut if the modal (any modal) is open.
-Create upload fetch on backend so the player on client side can now upload their profile pictures and render it to client side.
-Add a hash_profile on player info collection.
-Add delete image on album that is if player is guest that will be validated on backend.
-Add a partial update on modify data route, it is now working when player only change ign or description, but if it choose to update profile too then it will be reflect on db otherwise only the ign and description.

+++ Server/websocket_game_stuff.js
@@ -3,6 +3,32 @@ const accountModel = require("./accountMongooseSchema");
 const playerInfoModel = require("./playerInformationMongooseSchema");
 const broadcastSocket = require("./websocket_broadcast");
 
+const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
+
+require("dotenv").config({ path: require("path").resolve(__dirname, "../keys.env")})
+
+async function delete_image(profile_hash){
+    try{
+        if(profile_hash != "ajVzRmV"){
+            const deleteImg = await fetch(`https://api.imgur.com/3/image/${profile_hash}`, {
+                method: "DELETE",
+                headers: {
+                    Authorization: `Bearer ${process.env.IMGUR_ACCESS_TOKEN}`
+                }
+            });
+
+            const deleteImg_res = await deleteImg.json();
+
+            if(!deleteImg_res.success){
+                console.log(deleteImg_res)
+            }
+        }
+    }
+    catch(err){
+        console.log(err);
+    }
+}
+
 module.exports = (wss)=>{
     wss.on('connection', (ws) => {
         // Listen for messages from clients
@@ -119,7 +145,12 @@ async function deleteGuestPlayer_account(username) {
         let deleteAcc = await accountModel.findOneAndDelete({ username: username, account_type: "Guest" });
 
         if(deleteAcc){
-            await playerInfoModel.findOneAndDelete({ username: username });
+            const find_player = await playerInfoModel.findOne({ username: username })
+
+            if(find_player){
+                await delete_image(find_player.profile_hash);
+                await playerInfoModel.findOneAndDelete({ username: username });
+            }
         }
         else{
             await accountModel.findOneAndUpdate({ username: username }, { $set: { isOnline: false }, new: true })
